Fix problem with too long value for integer
git-svn-id: https://svn.l3s.uni-hannover.de/learnweb/trunk/InterWebJ@4621 422a2c9f-66af-4668-9993-60dc3b0871b8

Former-commit-id: a18fe8c0f0d887c3deceeed33216017b725a8496 [formerly 0a9d2f7befc62a1566920eb192f13fab48c7acf5]
Former-commit-id: 79f4a915989f63cb0a7dc970b1b545a61d817838

+++ connectors/BingAzureConnector/src/de/l3s/interwebj/connector/bingAzure/BingAzureConnector.java
@@ -225,9 +225,9 @@ public class BingAzureConnector extends AbstractServiceConnector
 		}
 		
 		Element rootProp = rootElement.element("entry").element("content").element("properties");
-		int webTotal = !rootProp.elementText("WebTotal").isEmpty() ? Integer.parseInt(rootProp.elementText("WebTotal")) : 0;
-		int imageTotal = !rootProp.elementText("ImageTotal").isEmpty() ? Integer.parseInt(rootProp.elementText("ImageTotal")) : 0;
-		int videoTotal = !rootProp.elementText("VideoTotal").isEmpty() ? Integer.parseInt(rootProp.elementText("VideoTotal")) : 0;
+		long webTotal = !rootProp.elementText("WebTotal").isEmpty() ? Long.parseLong(rootProp.elementText("WebTotal")) : 0;
+		long imageTotal = !rootProp.elementText("ImageTotal").isEmpty() ? Long.parseLong(rootProp.elementText("ImageTotal")) : 0;
+		long videoTotal = !rootProp.elementText("VideoTotal").isEmpty() ? Long.parseLong(rootProp.elementText("VideoTotal")) : 0;
 		queryResult.setTotalResultCount(webTotal + imageTotal + videoTotal);
 		
 		int index = 1;
